requirements.txt
# LunaClaw Brief — Dependencies
aiohttp>=3.9
requests>=2.31
jinja2>=3.1
pyyaml>=6.0
weasyprint>=60.0    # 可选，用于 HTML → PDF
